requirements.txt
aniso8601==9.0.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
